Working on breaking up handler logic to other files
